frontend: Move apiProxy to a separate folder
